refactor: add util to check if jwt is expired, add selector for derived RefreshToken
